azure_function/requirements.txt
azure-functions==1.21.3
azure-storage-file-datalake==12.19.0
azure-identity==1.21.0
openai>=1.50.0
requests==2.32.3
